Share driver code between perf and test
